Refactor specific_data.py (#64)
* Fix unit of molar heat

* Adjust command line parser and help messages

The adjustmenst reflect how the code as is actually works and the extended
help message give actual examples of how to use the command.

* Clarify help message

* Make command executable

* Add option to quit adding properties

* Refactor output to use functionality from argparse

Default behavior is now more robust for output. Changed write functions
to use less of global state.

* Parameterize help message fors default output file name

* Improve prompt to quit interactive loop

* Add example with properties and interactive

* Remove debug cruft

* Refactor write functions to not use global variables

* Skip already selected in interactive phase given by properites

* Improve help message

* Refactor logic for output in the default case

USe the functionality provided by argparse for default values of an
optional command line argument.  In this case --output.

* Restructure code by moving already defined functions up

* Refactor structure of code to use functions

This is s step towards further simplification of each step.

* Refactor use of keys and clear screen for interactive use only

* Add the main guard

* Use argparse hook to parse comma separated properties

This provides a list of arguments directly.  Find the set intersection
and difference to filter out correctly named properties.

Structure the code so as to keep the terminal output separate.

* Refactor interactive paring to us local functions

* Use result provided by default_input

* Use pylint to clean up

* Improve docstrings

* Format help message

* Refactor name and replace with namedtuple